CyberSentinel AI (3)/requirements.txt
# Core Framework Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database Dependencies
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0
pymongo==4.6.0
aiosqlite==0.19.0

# AI/ML Dependencies
scikit-learn==1.3.2
numpy==1.24.3
pandas==2.1.4

# Natural Language Processing
nltk==3.8.1
textblob==0.17.1

# Security and Cryptography
cryptography==41.0.8
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.6

# API and HTTP
requests==2.31.0
httpx==0.25.2
aiohttp==3.9.1

# Configuration and Environment
python-dotenv==1.0.0
pyyaml==6.0.1

# Logging and Monitoring
structlog==23.2.0
prometheus-client==0.19.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Code Quality
black==23.11.0
flake8==6.1.0
mypy==1.7.1
isort==5.12.0

# Development Tools
jupyter==1.0.0
ipython==8.18.1

# Email Processing
email-validator==2.1.0

# File Processing
python-magic==0.4.27
Pillow==10.1.0

# Async Support
asyncio-mqtt==0.16.1
redis==5.0.1

# Documentation
mkdocs==1.5.3
mkdocs-material==9.4.8

# Anomaly Detection
isolation-forest==0.1.0
pyod==1.1.0

# Model Management
mlflow==2.8.1
wandb==0.16.1
joblib==1.3.2

# New security dependencies
PyJWT==2.8.0
bcrypt==4.1.2

# New monitoring and metrics
psutil==5.9.6

# New caching
aioredis==2.0.1

# Enhanced Architecture Dependencies
# Ollama for local LLM inference
ollama==0.1.7

# LeetTools for RAG pipeline
leettools==0.1.0

# Atomic Agents for modular agent framework
atomic-agents==0.2.0

# Instructor for structured outputs
instructor==0.4.8

# FAISS for vector database
faiss-cpu==1.7.4

# Nomic embeddings
nomic==2.0.8

# PEFT for fine-tuning
peft==0.7.1
transformers==4.36.2
torch==2.1.1

# Vector database
chromadb==0.4.18

# Added from the code block
asyncio==3.4.3 